21f3001013/Backend/requirements.txt
alembic==1.11.1
amqp==5.1.1
aniso8601==9.0.1
Babel==2.12.1
bcrypt==4.0.1
#billiard==4.1.0
#blinker==1.6.2
#boto3==1.28.10
#botocore==1.31.10
cachelib==0.9.0
celery[redis]
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
##colorama==0.4.6
DateTime==5.2
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.2.3
Flask-BabelEx==0.9.4
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Reuploaded==1.3.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
flask-sse 
#gunicorn
#gevent
#greenlet==2.0.2
#idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
#jmespath==1.0.1
#kombu==5.3.1
#mailhog
#Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.1
passlib==1.7.4
peewee==3.16.0
pluggy==1.2.0
prompt-toolkit==3.0.39
pycryptodome==3.18.0
PyJWT==2.8.0
pytest==7.4.0
python-dateutil==2.8.2
#pytz==2022.7.1
PyYAML==6.0.1
redis==5.0.0
#s3transfer==0.6.1
#secret==0.8
#sendgrid
#six==1.16.0
#speaklater==1.3
SQLAlchemy==2.0.7
#tabulate==0.9.0
#typing_extensions==4.5.0
#tzdata==2023.3
urllib3==1.26.16
#vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
WTForms==3.0.1
#zope.interface==6.0
pyflakes